Moved rule validation to a separate extension class
